move PlayFriendOptions to separate file
