+++ imports/ui/pages/components/RightSidebar/PlayRightSidebar.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Meteor } from "meteor/meteor";
-import { Button, Form, InputNumber, Radio, Tabs } from "antd";
+import { Button, Tabs } from "antd";
 import KibitzChatApp from "./../Chat/KibitzChatApp";
 import PersonalChatApp from "./../Chat/PersonalChatApp";
 import GameHistory from "./elements/GameHistory";
@@ -8,157 +8,12 @@ import { GameControlBlock } from "./elements/GameControlBlock";
 
 import { Logger } from "../../../../../lib/client/Logger";
 import PlayChooseBot from "./PlayChooseBot";
-import { findRatingObject } from "../../../../../lib/ratinghelpers";
-import { DynamicRatingsCollection } from "../../../../api/client/collections";
-import { translate } from "../../../HOCs/translate";
 import PlayWithFriend from "./PlayWithFriend";
+import PlayFriendOptions from "./PlayFriendOptions";
 
 const log = new Logger("client/PlayRightSidebar");
 
 const { TabPane } = Tabs;
-
-class PlayFriendOptions extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      color: "random",
-      incrementOrDelayType: "inc",
-      initial: 7,
-      incrementOrDelay: 0,
-      ratingType: "none"
-    };
-  }
-
-  componentDidMount() {
-    this.updateRating();
-  }
-
-  handleChangeColor = e => {
-    this.setState({
-      color: e.target.value
-    });
-  };
-
-  handleChangeIncrementOrDelayType = e => {
-    this.setState({
-      incrementOrDelayType: e.target.value
-    });
-  };
-
-  handleChange = inputName => {
-    return number => {
-      const newState = {};
-
-      newState[inputName] = number;
-
-      this.setState(newState, () => {
-        this.updateRating();
-      });
-    };
-  };
-
-  updateRating = () => {
-    const { initial, incrementOrDelay, incrementOrDelayType } = this.state;
-
-    const ratingObject = findRatingObject(
-      0,
-      "white", // Right now white and black always match, so just hard code
-      initial,
-      incrementOrDelay,
-      incrementOrDelayType,
-      DynamicRatingsCollection.find().fetch()
-    );
-
-    this.setState({ ratingType: ratingObject.rating_type });
-  };
-
-  handlePlay = () => {
-    const { onPlay } = this.props;
-
-    let { color } = this.state;
-    const { ratingType, incrementOrDelayType, initial, incrementOrDelay } = this.state;
-
-    if (color === "random") {
-      color = Math.random() < 0.5 ? "white" : "black";
-    }
-    onPlay({
-      ratingType,
-      color,
-      incrementOrDelayType,
-      initial,
-      incrementOrDelay
-    });
-  };
-
-  render() {
-    const { onClose, translate } = this.props;
-    const { initial, incrementOrDelay, incrementOrDelayType, ratingType, color } = this.state;
-
-    return (
-      <div className="play-friend">
-        <div className="play-friend__head">
-          <h2 className="play-friend__name-title">{translate("createGame")}</h2>
-          <Button onClick={onClose}>{translate("BACK")}</Button>
-        </div>
-        <Form
-          className="play-bot__form"
-          layout="vertical"
-          initialValues={{
-            initial,
-            incrementOrDelay,
-            color: "random"
-          }}
-        >
-          <Form.Item label={translate("timeControl")} name="time-control">
-            <Radio.Group
-              onChange={this.handleChangeIncrementOrDelayType}
-              value={incrementOrDelayType}
-            >
-              <Radio.Button value="inc">{translate("control.inc")}</Radio.Button>
-              <Radio.Button value="none">{translate("control.none")}</Radio.Button>
-              <Radio.Button value="us">{translate("control.us")}</Radio.Button>
-              <Radio.Button value="bronstein">{translate("control.bronstein")}</Radio.Button>
-            </Radio.Group>
-            <div className="play-right-sidebar__inc-deley-wrap">
-              <Form.Item label={translate("initial")} name="initial">
-                <InputNumber
-                  min={0}
-                  disabled={incrementOrDelayType === "none"}
-                  value={initial}
-                  onChange={this.handleChange("initial")}
-                />
-              </Form.Item>
-              <Form.Item label={translate("incrementOrDelay")} name="incrementOrDelay">
-                <InputNumber
-                  min={0}
-                  disabled={incrementOrDelayType === "none"}
-                  value={incrementOrDelay}
-                  onChange={this.handleChange("incrementOrDelay")}
-                />
-              </Form.Item>
-            </div>
-          </Form.Item>
-          <Form.Item label={translate("ratingType")} name="ratingType">
-            <p>{translate(`ratings.${ratingType}`)}</p>
-          </Form.Item>
-          <Form.Item label={translate("color")} name="color">
-            <Radio.Group onChange={this.handleChangeColor} value={color}>
-              <Radio.Button value="random">{translate("colors.random")}</Radio.Button>
-              <Radio.Button value="white">{translate("colors.white")}</Radio.Button>
-              <Radio.Button value="black">{translate("colors.black")}</Radio.Button>
-            </Radio.Group>
-          </Form.Item>
-          <Button type="primary" onClick={this.handlePlay}>
-            {translate("selectOpponent")}
-          </Button>
-        </Form>
-      </div>
-    );
-  }
-}
-
-const EnhancedPlayFriendOptions = translate("Play.PlayFriendOptions")(PlayFriendOptions);
 
 class PlayBlock extends Component {
   constructor(props) {
@@ -236,7 +91,7 @@ class PlayBlock extends Component {
     }
     if (!isPlaying && this.state.status === "play-friend-options") {
       return (
-        <EnhancedPlayFriendOptions
+        <PlayFriendOptions
           onClose={() => {
             this.setState({ status: "none" });
           }}
